CDPE-2915: Support for executing windows jobs
Because we are using the orcestrator to execute bolt tasks via PCP, we
don't have to worry about windows vs *nix when executing our bolt plan.
However, we DO have to care about the system when we go to execute the
CD4PE job on the target machine. This commit adds handling for windows
by prefacing system commands with 'powershell' and wrapping them in some
absurd windows-specific characters that allows the command to execute
properly.