CDPE-3041: Extend default ReadTimout to 5 minutes
With this change, we increase the default net:http client read_timeout
to 5 minutes. This means that once a connection is established, and
a request has been made, the client will wait up to 5 minutes to get a
response before throwing.

We also added error handling in the event that this timeout is thrown,
and log a useful message to the user.