CDPE-3177: Set HOME and REPO_DIR env vars
Prior to this change, jobs run via orchestrator on a puppet agent would
not have their $HOME env var set, causing problems for many common
tasks run inside jobs, such as `bundle pdk install`. We also had no way
for the writer of the cd4pe job script to reference the control repo
they wished to run tests on, forcing them to use a relative path in
their job script to locate the repos dir.

With this change, we set the $HOME var on the puppet agent before
executing the job script so it is accessible for the duration of the
job, and the $REPO_DIR var to the parent of the control repo so the
cd4pe job script writer can reliably access the control repo.