CDPE-3177: Set $HOME var using ruby implementation
Prior to this change, we were setting the $HOME env var by shelling out
to the system, and relying on the `~` command to return the home dir.
While this works fine on certain systems, its not guaranteed to be
system-agnostic. Therefore, we replaced this with a ruby implementation,
which we can certain is present on any system executing this task.