CDPE-3250 Handle long Puppet agent job logs
With this change, we send the detailed log output of Puppet
agent jobs to a new endpoint in CD4PE.

Prior to this commit, if a Puppet agent job produced too many log
entries, the Orchestrator would crash and the job would fail.
See detail from ticket `ORCH-2255`...
```
Any user using bolt with PXP will likely come across this. Users
will not expect there to be limitations when uploading a 5Mb file
or run a 1500 line script, both if which will cause orchestrator/PXP
to fall over completely. Users will expect this to work as it
works fine over SSH or WinRM.
```