CDPE-3334: make module backwards compatible with CD4PE 3.9.3 and lower
Prior to this commit, we implemented a new feature to send CD4PE job
logs directly from the puppet agent to a new CD4PE endpoint to avoid
problems with the Puppet orchestrator. However, we failed to make these
changes backwards compatible, such that customers consuming the CD4PE
jobs module would see failed jobs if they failed to update CD4PE to
3.10.0.

With this change, we add logic to handle this case, and fallback to the
old behavior if the customer is using a version of CD4PE 3.9.3 or lower.