CDPE-3457: fix res code handling when saving logs
Prior to this change, when a cd4pe job sent logs to cd4pe, we
misinterpreted the response code, and also sent the logs to stdout.
This caused a bug where the job log would appear to have sent the logs
to cd4pe, thereby circumventing the need to send logs to stdout, but
would actually send logs to stdout on the sly, causing jobs with large
logs to hang.

With this change, we fix the response code handling to properly
interpret the response code from cd4pe.