CDPE-3033: Make http read timeout configurable
Prior to this change, the run_cd4pe_job had a default of 10 minutes when
hitting a CD4PE endpoint. This worked fine for most users, but a few
customers with large git repos ran into this timeout, and needed it to
be configurable.

With this change, we add an optional parameter to the task to allow the
caller to specify a custom net http read timeout in seconds, and then
use this parameter to set the connection timeout on the Net:HTTP client.

Because this parameter is optional, this change is backwards compatible
with existing versions of CD4PE, and can be released independently.