CDPE-3480: Update job to use registry creds if provided
Update the job task to accept a Docker config.json as a base64-encoded
string. If provided, this file will be used when pulling the specified
image to add custom registry credentials.

For each host with custom auth in the Docker config, add the custom CA
cert (if provided) to Docker's cert trust for that host.