CDPE-3648: Fix logging for failed http requests
Prior to this change, if an unknown http request failure occurred when
contacting CD4PE, an exception was thrown to the outer-most exception
handler. As a result, it was almost impossible to debug new, unexpected
http request failures.

With this change, we add a few improvements to make this easier in the
future:
 - Add exception handling for all unknown http request error, displaying
   the error in the job log before exiting the job script.
 - Time each request, and display the total request time in the job log
   (helps when determining the source of timeout errors)
 - Print the ruby stacktrace if the outmost exception handling clause is
   encountered.
 - Add exception handling when trying to write the cd4pe job payload to
   disk.