(CDPE-3886) Switch local registry cert from hardlink to symlink
Prior to this commit, if a user had a separate filesystem mounted
on their cd4pe_working_dir, an airgap installation would fail to
run jobs because it attempted to create a hardlink between that
filesystem and the ca in /etc.

With this commit, we switch the hardlink to a symlink, which allows
it to cross filesystem boundaries.