(CDPE-3377) Update request token logic to use proper auth
Prior to this commit, when making requests to CD4PE from the client, we
used a `Bearer token` prefix on our token. Since we modified CD4PE's
auth system to use token-based auth, we need to modify the way we do
token-based auth to adhere to our new standards.

With this change, we remove the `Bearer token ` prefix from token-based
auth to adhere to our current standards in CD4PE.

Note:
This change does not require a new release of cd4pe_jobs. CD4PE is
backwards compatible with the old code, so we can just release this
change whenever another release goes out.