(CDPE-3377) Fix unit tests and run them in CI on unix/windows
Prior to this commit, we introduced changes to the cd4pe_jobs module
without updating our unit tests. These tests were not run as part of any
CI process, so it makes sense that we missed them.

With this commit, we fix our tests and update them to run in CI via
Github Actions. As an added bonus, we can now validate our windows-based
jobs by utilizing the Github Actions windows test runner, so as part of
this commit, we modified our tests to run on both unix and windows.

Note:
The way I chose to differentiate between windows and unix code paths in
our test suite is less than ideal, but my goal here was simply to get
the tests running on both windows and unix in CI. We can make further
optimizations to this in the future, but since we plan to limit new
development on this module in favor of k8s-based jobs, it doesn't seem
worth the effort to further optimize this right now.