(maint) Update metadata to align with PDK 3.0.1
This commit adds the changes that the PDK made the metadata when
building the tarball for publishing via `pdk build`. Notably, it adds
some Puppet version requirements, representing the current supported
versions of Puppet. The Forge will map these to the corresponding PE
versions for display to the user. These version requirements are
non-binding, so it shouldn't negatively impact any users.

This commit also removes the outdated `pdk-template` lines from the
metadata.json. We don't want to make this module fully compatible with
modern PDK at the moment, but we also don't want to indicate that a very
old version was used to build it. From now on, devs should use `pdk
build --force` to build the module without allowing the PDK to make
changes to the metadata.