(CDPE-6808) Add podman support
This commit adds support for podman. Instead of just executing docker,
the task will look to see if either podman or docker are present, using
the first one found.

An environment variable is provided to give users a way to override the
automatic detection if there is a special case we haven't thought of.

Split out the runtime specific tests to separate specs, one for podman
and another for docker to allow them to run on github runners with the
appropriate runtime available.