(maint) Add logging when image pull fails
This commit adds logging to indicate that we fall back to assuming a
local version of the job image when pulling it from the remote fails.
This should help users recognize that the pull errors are red herrings
in air-gapped installs, and that the job will proceed, assuming the
image has been loaded into the runtime out of band.