(PE-42665) Make run_cd4pe_job task public
This commit undoes a previous change that made the `run_cd4pe_job` task
private. This hid the task from the PE console, which was desirable so
that users would not be able to run it directly, outside of CD. However,
with this change, users of PE 2025.6 were also no longer able to
give the CD service account user permission to run this task, meaning
that all CD jobs would fail. So until we figure out a better long-term
solution, this commit makes the task public again.

This commit also updates the task description to make it more clear how
it is intended to be run, and to discourage users from running it
directly.